feat: Improve styling of Formik example
