+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "react-responsive-select",
-  "version": "5.0.0-alpha-22",
+  "version": "5.0.0-alpha-23",
   "description": "A React customisable, touchable, single-select / multi-select form component. Built with keyboard and screen reader accessibility in mind.",
   "main": "dist/ReactResponsiveSelect.js",
   "types": "dist/ReactResponsiveSelect.d.ts",
