+++ src/ReactResponsiveSelect.css
@@ -20,7 +20,6 @@
   color: #555;
   position: relative;
   cursor: pointer;
-  height: 44px;
   line-height: 44px;
   background: #fff;
   border-radius: 4px;
