+++ src/stories/Info/Home/Home.stories.tsx
@@ -1,7 +1,7 @@
 import { storiesOf } from '@storybook/react';
 import * as React from 'react';
 import ReactResponsiveSelect from '../../../ReactResponsiveSelect';
-import { CaretIcon } from '../../components/CaretIcon';
+import { CaretIcon } from '../../components/Icons';
 import { MultiSelectOptionMarkup } from '../../components/MultiSelectOptionMarkup';
 import { ViewIframe } from '../../components/ViewIframe';
 import { FeaturesList } from './elements/FeaturesList';
