+++ src/stories/MultiSelect/noSelectionLabel-w-optHeader.stories.tsx
@@ -2,7 +2,7 @@ import { storiesOf } from '@storybook/react';
 import * as React from 'react';
 import ReactResponsiveSelect from '../../ReactResponsiveSelect';
 import { withStoryBookInfo } from '../../utils/withStoryBookInfo';
-import { CaretIcon } from '../components/CaretIcon';
+import { CaretIcon } from '../components/Icons';
 import { MultiSelectOptionMarkup } from '../components/MultiSelectOptionMarkup';
 
 import '../../ReactResponsiveSelect.css';
