+++ src/stories/Recipes/listening-for-rrs-changes.css
@@ -1,3 +1,5 @@
-.no-scroll-y {
-    overflow-y: hidden;
+@media screen and (max-width: 768px) {
+    .no-scroll-y {
+        overflow-y: hidden;
+    }
 }
