+++ src/stories/Recipes/using-formik.css
@@ -1,9 +1,23 @@
-.has-error .rrs__button {
-    border: 1px solid #f44336;
+.has-error {
+    background: #ffebee;
+    padding: 4px;
+}
+
+.has-error .rrs__label,
+.has-error .rss__button:focus .rrs__label,
+.has-error .rss__button:focus .rrs__label svg {
+    border: 1px solid #f44336!important;
+    color: #f44336;
+}
+
+.has-error .rrs__label svg * {
+    fill: #f44336;
 }
 
 .field-error-message {
     color: #f44336;
     font-size: 12px;
-    padding: 0.25rem;
+    padding: 0.5rem;
+    display: flex;
+    align-items: center;
 }
