+++ src/stories/stories.css
@@ -19,6 +19,38 @@ body {
   padding: 1rem;
 }
 
+.icon {
+  fill: currentColor;
+}
+
+button {
+  padding: 0.5rem 1rem;
+  border-style: none;
+  border-radius: 4px;
+  background-color: #2196F3;
+  font-size: 0.85rem;
+  color: #fff;
+  cursor: pointer;
+  outline: none;
+  -webkit-appearance: none;
+  box-shadow: 0 1px 2px 0 rgba(0, 0, 0, 0.2);
+}
+
+button:disabled {
+  opacity: 0.5;
+  cursor: not-allowed !important;
+}
+
+button + button {
+  margin-left: 0.5rem;
+}
+
+button.outline {
+  background-color: #eee;
+  border: 1px solid #aaa;
+  color: #555;
+}
+
 .badge {
   background: #f1f1f1;
   padding: 0.25rem 1rem;
@@ -113,5 +145,5 @@ video {
 }
 
 form > * {
-    margin: 1rem;
+    margin: 1rem 1rem 0 1rem;
 }
